Exclude content size from head request.
Servers are not required to send content-length for head
request. It may happen for example in case if Accept-Encoding is set to gzip
(this is a default in http-client) and server decided that it will gzip
the file.

So solution is to create new type 'HeadInfo' that will not keep 'hiSize'
field that is not reliable.